Refactor opensearch indices initialization
